define layout route for balances
