+++ deployables/app/app/routes/tokens/balances/layout.tsx
@@ -1,0 +1,14 @@
+import { Page } from '@/components'
+import { Outlet } from 'react-router'
+
+const BalancesLayout = () => (
+  <Page fullWidth>
+    <Page.Header>Balances</Page.Header>
+
+    <Page.Main>
+      <Outlet />
+    </Page.Main>
+  </Page>
+)
+
+export default BalancesLayout
